adding the array_to_list function, fixing indexes from int to long

+++ src/list_count.c
@@ -7,7 +7,7 @@
 
 #include "list.h"
 
-int list_count(list_t const *list)
+long list_count(list_t const *list)
 {
     return list->count;
 }
