+++ tests/tests_push_and_insert.c
@@ -22,7 +22,7 @@ Test(list_push_back, test_impl)
 
     cr_assert(eq(int, 0, list_push_back(list, strdup("Hello, World !"))));
     cr_assert(eq(int, 0, list_push_back(list, strdup("This is my tail !"))));
-    cr_assert(eq(int, 2, list_count(list)));
+    cr_assert(eq(long, 2, list_count(list)));
     cr_assert(eq(str, "Hello, World !", list->elems_head->elem));
     cr_assert(eq(str, "This is my tail !", list->elems_tail->elem));
     list_destroy(list);
@@ -34,7 +34,7 @@ Test(list_push_front, test_impl)
 
     cr_assert(eq(int, 0, list_push_front(list, strdup("Hello, World !"))));
     cr_assert(eq(int, 0, list_push_front(list, strdup("This is my head !"))));
-    cr_assert(eq(int, 2, list_count(list)));
+    cr_assert(eq(long, 2, list_count(list)));
     cr_assert(eq(str, "This is my head !", list->elems_head->elem));
     cr_assert(eq(str, "Hello, World !", list->elems_tail->elem));
     list_destroy(list);
